Export `IdTokenValidationConfig`
