+++ auth0_flutter_platform_interface/lib/auth0_flutter_platform_interface.dart
@@ -16,6 +16,7 @@ export 'src/request/request.dart';
 export 'src/request/request_options.dart';
 export 'src/user_agent.dart';
 export 'src/user_profile.dart';
+export 'src/web-auth/id_token_validation_config.dart';
 export 'src/web-auth/web_auth_exception.dart';
 export 'src/web-auth/web_auth_login_input.dart';
 export 'src/web-auth/web_auth_logout_input.dart';
