+++ auth0_flutter_platform_interface/lib/src/web-auth/id_token_validation_config.dart
@@ -1,0 +1,7 @@
+class IdTokenValidationConfig {
+  final int? leeway;
+  final String? issuer;
+  final int? maxAge;
+
+  const IdTokenValidationConfig({this.leeway, this.issuer, this.maxAge});
+}
